rename files Signed-off-by: Naman Anand <naman.anand@nimbleedgehq.ai>
