+++ agents/examples/android/app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
         </activity>
 
         <service
-            android:name="ai.nimbleedge.notifications_summarizer.impl.notification.NotificationListener"
+            android:name="dev.deliteai.notifications_summarizer.impl.notification.NotificationListener"
             android:label="Notification Listener"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE"
             android:exported="true">
@@ -45,17 +45,17 @@
         </service>
 
         <service
-            android:name="ai.nimbleedge.notifications_summarizer.impl.services.SummarizationForegroundService"
+            android:name="dev.deliteai.notifications_summarizer.impl.services.SummarizationForegroundService"
             android:enabled="true"
             android:exported="false"
             android:foregroundServiceType="dataSync" />
 
         <receiver
-            android:name="ai.nimbleedge.notifications_summarizer.impl.broadcast.AlarmBroadcastReceiver"
+            android:name="dev.deliteai.notifications_summarizer.impl.broadcast.AlarmBroadcastReceiver"
             android:enabled="true"
             android:exported="false">
             <intent-filter android:priority="1000">
-                <action android:name="ai.nimbleedge.notifications_summarizer.ALARM_TRIGGER" />
+                <action android:name="dev.deliteai.notifications_summarizer.ALARM_TRIGGER" />
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
             </intent-filter>
         </receiver>
