+++ agents/notifications_summarizer/android/src/main/java/dev/deliteai/notifications_summarizer/dataModels/NotificationSummarizerConfig.kt
@@ -1,0 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package dev.deliteai.notifications_summarizer.dataModels
+
+data class NotificationSummarizerConfig(
+    val onScheduledSummaryReady: (notificationSummary: NotificationSummary?) -> Unit,
+)
