+++ agents/notifications_summarizer/android/src/main/java/dev/deliteai/notifications_summarizer/impl/data/local/Converters.kt
@@ -1,0 +1,18 @@
+/*
+ * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package dev.deliteai.notifications_summarizer.impl.data.local
+
+import androidx.room.TypeConverter
+import java.time.LocalDate
+
+class Converters {
+    @TypeConverter
+    fun fromLocalDate(date: LocalDate): String = date.toString()
+
+    @TypeConverter
+    fun toLocalDate(value: String): LocalDate = LocalDate.parse(value)
+}
