+++ agents/notifications_summarizer/android/src/main/java/dev/deliteai/notifications_summarizer/impl/services/SummarizationForegroundService.kt
@@ -1,0 +1,104 @@
+/*
+ * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package dev.deliteai.notifications_summarizer.impl.services
+
+import dev.deliteai.notifications_summarizer.dataModels.NotificationSummarizerConfig
+import dev.deliteai.notifications_summarizer.impl.DependencyContainer
+import dev.deliteai.notifications_summarizer.impl.LlmManager
+import dev.deliteai.notifications_summarizer.impl.common.Constants.TAG
+import dev.deliteai.notifications_summarizer.impl.data.local.NotificationSummaryDao
+import dev.deliteai.notifications_summarizer.impl.data.local.toNotificationSummaryEntity
+import dev.deliteai.notifications_summarizer.impl.notification.NotificationListener
+import dev.deliteai.notifications_summarizer.impl.notification.NotificationManager
+import android.app.Service
+import android.content.Intent
+import android.os.IBinder
+import android.util.Log
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
+
+class SummarizationForegroundService : Service() {
+    private val serviceScope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
+    private var summarizeJob: Job? = null
+    private lateinit var llmManager: LlmManager
+    private lateinit var notificationManager: NotificationManager
+    private lateinit var notificationSummaryDao: NotificationSummaryDao
+    private lateinit var config: NotificationSummarizerConfig
+
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        val dependencyContainer = DependencyContainer.getInstance()
+
+        notificationManager = dependencyContainer.getNotificationManager()
+        notificationSummaryDao = dependencyContainer.getNotificationSummaryDao()
+        config = dependencyContainer.getConfig()
+        llmManager = dependencyContainer.getLlmManager()
+
+        //we need only one notification for multiple services
+        val foregroundNotificationId = 746353
+        //TODO: discuss the copy of notification
+
+        startForeground(
+            foregroundNotificationId, notificationManager.buildForegroundNotification(
+                title = "Notification Summarization",
+                message = "Summarization in progress",
+            )
+        )
+
+
+        summarizeJob?.cancel()
+        summarizeJob = serviceScope.launch {
+            summarizeNotifications()
+            if (!serviceScope.isActive) stopSelf()
+        }
+        return START_REDELIVER_INTENT
+    }
+
+    override fun onBind(intent: Intent?): IBinder? = null
+
+    private suspend fun summarizeNotifications() {
+        waitForNotificationListener()
+
+        val notifications = NotificationListener.getNotifications()
+        if (notifications.isEmpty()) return
+
+        val summary = llmManager.getSummary(notifications)
+
+        notificationSummaryDao.insert(
+            summary.toNotificationSummaryEntity()
+        )
+
+        config.onScheduledSummaryReady(summary)
+    }
+
+    private suspend fun waitForNotificationListener() {
+        val notificationListenerTimeoutMs = 120000
+        val pollFrequencyMs = 1000L
+        var timeMs = 0L
+
+        while (!NotificationListener.isListenerConnected()) {
+            if (timeMs > notificationListenerTimeoutMs) {
+                Log.e(TAG, "summarizeNotifications: Notification Listener Timeout")
+                return
+            }
+
+            delay(pollFrequencyMs)
+            timeMs += pollFrequencyMs
+        }
+    }
+
+    override fun onDestroy() {
+        summarizeJob?.cancel()
+        serviceScope.cancel()
+        super.onDestroy()
+    }
+}
